Relocated tests from recipe_interpreters to the test folder
