+++ src/openag_brain/recipe_interpreters.py
@@ -145,82 +145,3 @@ def calc_phase_and_time_remaining(duration_of_phases_steps, current_time, start_
             current_phase_number = i
             break
     return current_phase_number, duration_in_phase
-
-
-def test_calc_phase_and_time_remaining():
-    duration_of_phases_steps = [(336, 24), (480, 24), (168, 24)]
-    start_time = datetime.strptime("2017-04-17 14:00", "%Y-%m-%d %H:%S")
-    current_time = datetime.strptime("2017-04-18 20:00", "%Y-%m-%d %H:%S")
-    current_phase_number, duration_in_step = calc_phase_and_time_remaining(duration_of_phases_steps,
-                                           current_time, start_time)
-    assert (current_phase_number, duration_in_step) == (0, 6)
-
-    start_time = datetime.strptime("2017-04-17 14:00", "%Y-%m-%d %H:%S")
-    current_time = datetime.strptime("2017-04-18 03:00", "%Y-%m-%d %H:%S")
-    current_phase_number, duration_in_step = calc_phase_and_time_remaining(duration_of_phases_steps,
-                                           current_time, start_time)
-    print(current_phase_number, duration_in_step)
-    assert (current_phase_number, duration_in_step) == (0, 13)
-
-
-def test_calculate_max_duration_from_step():
-    step = { "air_temperature": [{"start_time": 0, "end_time": 6, "value": 20},
-                                  {"start_time": 6, "end_time": 18, "value": 23},
-                                  {"start_time": 18, "end_time": 24, "value": 19}],
-              "nutrient_flora_duo_a": [{"start_time": 0, "end_time": 6, "value": 5},
-                                       {"start_time": 6, "end_time": 18, "value": 2},
-                                       {"start_time": 18, "end_time": 24, "value": 5}],
-              "nutrient_flora_duo_b": [{"start_time": 0, "end_time": 6, "value": 2}],
-              "light_illuminance": [{"start_time": 0, "end_time": 6, "value": 4},
-                                    {"start_time": 18, "end_time": 24, "value": 3}],
-        }
-    max_time = calculate_max_duration_from_step(step)
-    assert max_time == 24
-
-
-def test_calc_duration_of_phases_steps():
-    phases = [    # Previously operations,  Needs to be an ordered dictionary.
-          { "name": "early",
-            "cycles": 14,    # Add check for duration of a step to be a total of 24 hours. (Not a necessarity but valuable for consistency/simplicity)
-            "step": { "air_temperature": [{"start_time": 0, "end_time": 6, "value": 20},
-                                          {"start_time": 6, "end_time": 18, "value": 23},
-                                          {"start_time": 18, "end_time": 24, "value": 19}],
-                      "nutrient_flora_duo_a": [{"start_time": 0, "end_time": 6, "value": 5},
-                                               {"start_time": 6, "end_time": 18, "value": 2},
-                                               {"start_time": 18, "end_time": 24, "value": 5}],
-                      "nutrient_flora_duo_b": [{"start_time": 0, "end_time": 6, "value": 2}],
-                      "light_illuminance": [{"start_time": 0, "end_time": 6, "value": 4},
-                                            {"start_time": 18, "end_time": 24, "value": 3}],
-                },
-          },
-          { "name": "middle",
-            "cycles": 20,
-            "step": { "air_temperature": [{"start_time": 0, "end_time": 6, "value": 20},
-                                          {"start_time": 6, "end_time": 18, "value": 23},
-                                          {"start_time": 18, "end_time": 24, "value": 19}],
-                      "nutrient_flora_duo_a": [{"start_time": 6, "end_time": 18, "value": 1.4}],
-                      "nutrient_flora_duo_b": [{"start_time": 6, "end_time": 18, "value": 0.7}],
-                      "light_illuminance": [{"start_time": 6, "end_time": 18, "value": 1}],
-                },
-          },
-          { "name": "late",
-            "cycles": 7,
-            "step": { "air_temperature": [{"start_time": 6, "end_time": 24, "value": 23}],
-                    },
-          }
-        ]
-    duration_of_phases_steps = calc_duration_of_phases_steps(phases)
-    assert duration_of_phases_steps == [(336, 24), (480, 24), (168, 24)]
-
-
-def test_determine_value_for_step():
-    variable_step_data = [{"start_time": 0, "end_time": 6, "value": 20},
-                          {"start_time": 6, "end_time": 18, "value": 23},
-                          {"start_time": 18, "end_time": 24, "value": 19}]
-    duration_in_step = 15
-    value = determine_value_for_step(variable_step_data, duration_in_step)
-    assert value == 23
-    duration_in_step = 0
-    value = determine_value_for_step(variable_step_data, duration_in_step)
-    assert value == 20
-
